refactor : postPoint의 타입 Long으로 변경

+++ src/main/java/com/kakaotech/team14backend/outer/post/dto/GetPopularPostDTO.java
@@ -2,5 +2,5 @@ package com.kakaotech.team14backend.outer.post.dto;
 
 import java.util.List;
 
-public record GetPopularPostDTO(Long postId, String imageUri, List<String> hashTags, Long likeCount, Integer postPoint, String nickname, Integer postLevel) {
+public record GetPopularPostDTO(Long postId, String imageUri, List<String> hashTags, Long likeCount, Long postPoint, String nickname, Integer postLevel) {
 }
